Backend/requirements.txt
#requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
oracledb==1.4.2
mysql-connector-python==8.2.0
ibm-db==3.2.6
tabulate==0.9.0
typing-extensions==4.8.0
pyodbc==4.0.39
pydantic==1.10.13
sse-starlette>=1.3.2
psutil==5.9.8
requests==2.32.3
httpx==0.27.2
ibm-watsonx-ai==1.3.34